pyproject.toml
[project]
name = "o1-ai-lab"
version = "0.1.0"
description = "LangGraph-inspired autonomous agent system driven by OpenAI o1"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.30.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ai-lab"]

